Move JsCloseReason to api::room
