+++ jason/src/api/mod.rs
@@ -89,11 +89,8 @@ impl Jason {
     /// streams etc.) respectively. All objects related to this [`Jason`] API
     /// object will be detached (you will still hold them, but unable to use).
     pub fn dispose(self) {
-        self.0.borrow_mut()
-            .rooms
-            .drain(..)
-            .for_each(|room| {
-                room.close(CloseByClientReason::RoomClosed.into());
-            });
+        self.0.borrow_mut().rooms.drain(..).for_each(|room| {
+            room.close(CloseByClientReason::RoomClosed.into());
+        });
     }
 }
