+++ jason/src/api/room.rs
@@ -23,11 +23,74 @@ use crate::{
         MediaStream, MediaStreamHandle, PeerEvent, PeerEventHandler,
         PeerRepository,
     },
-    rpc::{CloseByClientReason, CloseReason, JsCloseReason, RpcClient},
+    rpc::{CloseByClientReason, CloseReason, RpcClient},
     utils::Callback,
 };
 
 use super::{connection::Connection, ConnectionHandle};
+
+/// Reason of why Jason was closed.
+///
+/// This struct will be provided into `on_close_by_server` JS side callback.
+#[wasm_bindgen]
+pub struct JsCloseReason {
+    /// Is closed by server?
+    ///
+    /// `true` if [`CloseReason::ByServer`].
+    is_closed_by_server: bool,
+
+    /// Reason of closing.
+    reason: String,
+
+    /// Is closing considered as error.
+    ///
+    /// This field may be `true` only on closing by client.
+    is_err: bool,
+}
+
+impl JsCloseReason {
+    /// Creates new [`ClosedByServerReason`] with provided [`CloseReason`]
+    /// converted into [`String`].
+    ///
+    /// `is_err` may be `true` only on closing by client.
+    ///
+    /// `is_closed_by_server` is `true` on [`CloseReason::ByServer`].
+    pub fn new(reason: &CloseReason) -> Self {
+        match reason {
+            CloseReason::ByServer(reason) => Self {
+                reason: reason.to_string(),
+                is_closed_by_server: true,
+                is_err: false,
+            },
+            CloseReason::ByClient { reason, is_err } => Self {
+                reason: reason.to_string(),
+                is_closed_by_server: false,
+                is_err: *is_err,
+            },
+        }
+    }
+}
+
+#[wasm_bindgen]
+impl JsCloseReason {
+    /// `wasm_bindgen` getter for `reason` field.
+    #[wasm_bindgen(getter)]
+    pub fn reason(&self) -> String {
+        self.reason.clone()
+    }
+
+    /// `wasm_bindgen` getter for `is_closed_by_server` field.
+    #[wasm_bindgen(getter)]
+    pub fn is_closed_by_server(&self) -> bool {
+        self.is_closed_by_server
+    }
+
+    /// `wasm_bindgen` getter for `is_err` field.
+    #[wasm_bindgen(getter)]
+    pub fn is_err(&self) -> bool {
+        self.is_err
+    }
+}
 
 /// JS side handle to `Room` where all the media happens.
 ///
